create standalone version to stay compatible with old docs
